fix formatting

+++ demos/brownie/mixes/chainlink/contracts/PriceFeedConsumer.sol
@@ -1,11 +1,9 @@
-
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
 import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 
 contract PriceFeedConsumer {
-
     AggregatorV3Interface internal priceFeed;
 
     /**
@@ -25,12 +23,12 @@ contract PriceFeedConsumer {
     /**
      * Returns the latest price
      */
-    function getLatestPrice() public view returns (int) {
+    function getLatestPrice() public view returns (int256) {
         (
-            uint80 roundID, 
-            int price,
-            uint startedAt,
-            uint timeStamp,
+            uint80 roundID,
+            int256 price,
+            uint256 startedAt,
+            uint256 timeStamp,
             uint80 answeredInRound
         ) = priceFeed.latestRoundData();
         return price;
